Lie to nix about kvm support
NixOS test currently requires the "kvm" system feature flag, but the
GitHub actions runners don't enable kvm support. Interestingly the nixos
tests still run on systems without kvm support, but they do run slower.
As this kvm support is not necessary and saying that it is available
when it is not does not seem to cause any adverse effects, this is a
tolerable solution at this time.

For more context see: https://github.com/NixOS/nix/issues/8316